requirements.txt
django==5.1.6
mysqlclient==2.2.0
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
